Progress upload
Ironed out some kinks with HP tracking in screen_capture. Just messing around with keybaordComos and windowsInputHandler. Beginning to put down the base code for the agent file. In the future the code can be separated into different files. For now just adding all the classes into this one file.