Updated getCurrentScreen method of Vision class
Added image downsizing so that the returned numpy array is a gray version of the image at a resolution of 85 x 85. I'm using the Lanczos algorithm because it is the fastest algorithm and with the inital image being so small the speed should not have a huge impact on the result.