About to run the test with the updated code.
The model can now choose from 84 different inputs of button combinations. The memory capacity was doubled. I may need to alter the learning rate and loss function to reflect the doubled memory. This test needs to be compared against a test that instead of doubling the memory, takes double the amount of frames to analyze.