Quality of life updates and images
Added functions to handle the manual exiting to character select and restarting matches. This way I don't have to start the trials at times that are convenient for when I wake up and get home from work. The bot can handle everything it self for the duration of the test. Added some screenshots for the report to illustrate how the bot looks and processes images.